ENH: Enable doxygen structs to be utilized in the breathelink
Currently, only doxygen classes are enabled which throws warnings for ITK Struct's defined in the ITKExamples. This will define a new Directive with the name breathelinkstruct, that will define a doxygenstruct instead of a doxygenclass. Thus, in the method run() when the Directive is processed we define it to be a struct if the Directive's name is breathelinkstruct.

In addition, the version for this respository was bumped from 0.1 to 0.2.

Lastly, the README was updated to describe the new Directive.